Implement read strategy for CBC
