--- src/otoole/results/__init__.py
@@ -1,3 +1,0 @@
-from otoole.results.convert import convert_cbc_to_csv
-
-__all__ = ["convert_cbc_to_csv"]
